refactor(main): extract process error handlers into lib/processErrorHandlers
Amp-Thread-ID: https://ampcode.com/threads/T-019e1fcf-7754-703e-9441-11cc2816c591
