+++ src/main/lib/processErrorHandlers.ts
@@ -1,0 +1,107 @@
+import { app } from 'electron'
+import * as mainTelemetry from './telemetry'
+import { _broadcastToRenderer } from './ipc/shared'
+import type { DatadogForwardedError } from '../../types/ipc'
+import { scrubAll } from '../../shared/piiScrub'
+
+/**
+ * Main-process error funnel: scrub, fan out to renderer for Datadog RUM,
+ * and capture in PostHog Node so nothing is lost when no panel is open.
+ */
+
+let processErrorHandlersRegistered = false
+
+function serializeUnknownError(error: unknown): { message: string; stack?: string } {
+  if (error instanceof Error) {
+    return {
+      message: error.message || error.name || 'Error',
+      stack: error.stack,
+    }
+  }
+  if (typeof error === 'string') {
+    return { message: error }
+  }
+  if (error === null || error === undefined) {
+    return { message: 'Unknown error' }
+  }
+  try {
+    return { message: JSON.stringify(error) }
+  } catch {
+    return { message: String(error) }
+  }
+}
+
+export function forwardDatadogError(payload: DatadogForwardedError): void {
+  const scrubbed: DatadogForwardedError = {
+    ...payload,
+    message: scrubAll(payload.message),
+    stack: payload.stack ? scrubAll(payload.stack) : undefined,
+    // Mark this error as already captured by main-process PostHog so the
+    // renderer's `onDatadogError` listener routes it to Datadog only and
+    // we don't double-count exceptions in PostHog.
+    skipPostHog: true,
+  }
+  // Broadcast to any open panel renderer so its `onDatadogError`
+  // listener can forward the error to Datadog RUM (the panel
+  // renderer hosts the telemetry bootstrap). When no panel is open
+  // the broadcast is a no-op and we still capture below via PostHog
+  // Node.
+  try {
+    _broadcastToRenderer('dd-error', scrubbed)
+  } catch {}
+  // Also surface to PostHog Node so we don't lose the error if no renderer is
+  // listening (render-process-gone, before-quit shutdown, no panel open yet).
+  try {
+    const err = new Error(scrubbed.message)
+    if (scrubbed.stack) err.stack = scrubbed.stack
+    mainTelemetry.captureException(err, {
+      origin: 'main-process',
+      source: scrubbed.source,
+      level: scrubbed.level ?? null,
+    })
+  } catch {}
+}
+
+export function registerProcessErrorHandlers(): void {
+  if (processErrorHandlersRegistered) return
+  processErrorHandlersRegistered = true
+
+  process.on('uncaughtExceptionMonitor', (error) => {
+    const serialized = serializeUnknownError(error)
+    forwardDatadogError({
+      source: 'main-uncaught-exception',
+      message: serialized.message,
+      stack: serialized.stack,
+      level: 'critical',
+      context: { origin: 'main-process' },
+    })
+  })
+
+  process.on('unhandledRejection', (reason) => {
+    const serialized = serializeUnknownError(reason)
+    forwardDatadogError({
+      source: 'main-unhandled-rejection',
+      message: serialized.message,
+      stack: serialized.stack,
+      level: 'error',
+      context: { origin: 'main-process' },
+    })
+  })
+
+  app.on('child-process-gone', (_event, details) => {
+    const extra = details as unknown as Record<string, unknown>
+    forwardDatadogError({
+      source: 'main-child-process-gone',
+      message: `Child process ${details.type} exited: ${details.reason}`,
+      level: 'error',
+      context: {
+        origin: 'main-process',
+        type: details.type,
+        reason: details.reason,
+        exitCode: details.exitCode,
+        name: extra['name'],
+        serviceName: extra['serviceName'],
+      },
+    })
+  })
+}
